Added option G3_LOGGER_NOT_INSTALLED
- Added option G3_LOGGER_NOT_INSTALLED
If setting the new option to ON, g3logger can be build within the parent project without need of being built and installed previously.
If the option is not provided the behaviour will not change. If the g3logger will be build with another version and another name, the name of the new lib can be provided by setting the G3LOG_LIBRARY within the cmake process.